\name{AggFPCTrBR}
\alias{AggFPCTrBR}
\title{Temporal aggregation for PFTs: get TrBR from vector of all PFTs, average over years}
\description{This function can be provided to \code{\link{IntegrationDataset}} to aggregate model results to the temporal resolution of the observations.}
\usage{AggFPCTrBR(x, agg)}
\arguments{
  \item{x}{full time series}
  \item{agg}{vector of grouping elements (years)}
}
\details{No details.}
\value{The function returns a the aggregated result.}
\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{IntegrationDataset}}}

