\name{AggMSC}
\alias{AggMSC}
\title{Temporal aggregation: mean seasonal cycle}
\description{This function can be provided to \code{\link{IntegrationDataset}} to aggregate model results to the temporal resolution of the observations.}
\usage{AggMSC(x, agg)}
\arguments{
  \item{x}{full time series}
  \item{agg}{vector of grouping elements (years)}
}
\details{No details.}
\value{The function returns the mean seasonal cycle}
\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{IntegrationDataset}}}

