\name{CheckLPJpar}
\alias{CheckLPJpar}
\title{Checks LPJ parameters 'LPJpar'}
\description{The function checks if LPJ parameters are within the lower and upper boundaries or are 0. }
\usage{CheckLPJpar(lpjpar, correct = FALSE)}
\arguments{
  \item{lpjpar}{object of class 'LPJpar'}
  \item{correct}{correct parameter values (TRUE) or return error message?}
}
\details{No details.}
\value{the function return an object of class 'LPJpar'}
\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{LPJpar}}		}

