\name{CheckMemoryUsage}
\alias{CheckMemoryUsage}
\title{Check usage of memory by R objects}
\description{Prints a message about the used memory and writes a file with the used memory per each R object.}
\usage{CheckMemoryUsage(...)}
\arguments{
  \item{\dots}{The function has no arguments.}
}

\value{a data.frame }

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





