\name{CombineLPJpar}
\alias{CombineLPJpar}
\title{Combines several 'LPJpar' objects into one}
\description{The function takes several lpjpar objects and combines them into one LPJpar object}
\usage{CombineLPJpar(lpjpar.l)}
\arguments{
  \item{lpjpar.l}{a list of \code{\link{LPJpar}} objects}
}
\details{No details.}
\value{The function returns a list of class 'LPJparList'}
\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{LPJpar}}}

