\name{CorrelationMatrixS}
\alias{CorrelationMatrixS}
\title{plot a correlation matrix}

\usage{CorrelationMatrixS(data, method = "spearman", iscor = NULL, main = "", 
    ...)}
\arguments{
  \item{data}{a correlation matrix or a data.frame (}
  \item{method}{method to compute the correlation}
  \item{iscor}{Is 'data' a correlation matrix?}
  \item{main}{main title for the plot}
  \item{\dots}{further arguments for \code{\link{plot}}}
}
\details{No details.}

\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{CombineRescueFiles}}}

