\name{EstOptimUse}
\alias{EstOptimUse}
\title{Estimate optimal number of jobs given a number of cluster nodes}

\usage{EstOptimUse(nodes = 16, wish = 1000)}
\arguments{
  \item{nodes}{number of cluster nodes that you want to use}
  \item{wish}{approx. number of elements}
}
\details{No details.}
\value{an integer value}
\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{WriteCLM}}}

