\name{InfoCLM}
\alias{InfoCLM}
\title{Returns information about a CLM file}

\usage{InfoCLM(file.clm, endian = "little", ...)}
\arguments{
  \item{file.clm}{CLM file name with extension *.clm}
  \item{endian}{endianess of the file}
  \item{\dots}{Further arguments (currently not used).}
}
\details{No details.}

\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{WriteLPJinput}}}

