\name{InfoNCDF}
\alias{InfoNCDF}
\title{Get information about variables in a NetCDF}

\usage{InfoNCDF(file)}
\arguments{
  \item{file}{file name}
}

\value{The function returns a list with information about the dimensions and variables in the NetCDF file.}

\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{WriteNCDF4}}}

