\name{LPJmLmdi-package}
\alias{LPJmLmdi-package}
\alias{LPJmLmdi}
\docType{package}
\title{Model-Data Integration for the LPJmL Dynamic Global Vegetation Model}
}
\description{Model-data integration framework for the LPJmL dynamic global vegetation model. Specifically, the package provides functions 1) to optimize LPJmL model parameters using the GENOUD genetic optimization algorithm, 2) to read and write LPJmL input data, and 3) to read LPJmL output files.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LPJmLmdi}
\packageIndices{LPJmLmdi}
}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}

Maintainer: \packageMaintainer{LPJmLmdi}
}

\keyword{ package }


