\name{LPJparList}
\alias{LPJparList}
\title{Create a list of 'LPJpar' objects}
\description{The function creates a list of \code{\link{LPJpar}} objects that can be used to compare parameters from different optimization experiments}
\usage{LPJparList(...)}
\arguments{
  \item{\dots}{objects of class \code{\link{LPJpar}}}
}
\details{No details.}
\value{The function returns a list of class 'LPJparList'}
\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{LPJpar}}}

