\name{LPJppNBP}
\alias{LPJppNBP}
\title{Post-process LPJmL model output: calculate NEE}
\description{The function calculates NEE from LPJmL model output}
\usage{LPJppNBP(path, start = 1982, end = 2011, sim.start.year = 1901, 
    ...)}
\arguments{
  \item{path}{directory with LPJmL outputs in *.bin format}
  \item{start}{first year for which the data should be converted to NetCDF}
  \item{end}{last year for which the data should be converted to NetCDF}
  \item{sim.start.year}{first year of the simulation}
  \item{\dots}{further arguments (currently not used)}
}


\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





