\name{plot.LPJparList}
\alias{plot.LPJparList}
\title{Plots to compare LPJpar objects}
\description{The function takes a \code{\link{LPJparList}} object and creates a plot to compare optimized parameters}
\usage{\method{plot}{LPJparList}(x, par.name = NULL, ...)}
\arguments{
  \item{x}{object of class 'LPJparList'}
  \item{par.name}{name(s) of the parameters that should be plotted}
  \item{\dots}{further arguments to \code{\link{plot.LPJpar}}	}
}
\details{No details.}

\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{LPJpar}}, \code{\link{plot.LPJpar}}	}

