\name{PlotWorld110}
\alias{PlotWorld110}
\title{Plot a world map based on 1:110Mio data}

\usage{PlotWorld110(admin = FALSE, lakes = TRUE, rivers = TRUE, col = c("black", 
    "blue", "red"), bg = NA, ...)}
\arguments{
  \item{admin}{Plot administrative boundaries?}
  \item{lakes}{Plot lakes?}
  \item{rivers}{Plot rivers?}
  \item{col}{Colors for (1) coastlines, (2) rivers and (3) administrative boundaries}
  \item{bg}{background color, default: NA (no background)}
  \item{\dots}{additional arguments to plot}
}



\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}




\examples{

PlotWorld110()

}
