\name{PrepareRestartFiles}
\alias{PrepareRestartFiles}
\title{Prepare restart files to restart OptimizeLPJgenoud}
\description{The function prepares all files that are needed to restart OptimizeLPJgenoud}
\usage{PrepareRestartFiles(file.optsetup, ...)}
\arguments{
  \item{file.optsetup}{OptimizeLPJgenoud setup file, ends with "_optsetup.RData"}
  \item{\dots}{further arguments (currently not used)}
}
\details{No details.}

\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{LPJfiles}}	}

