\name{ReadGrid}
\alias{ReadGrid}
\title{Reads a binary input data grid file.}

\usage{ReadGrid(file.grid = "cru.grid", endian = "little", ...)}
\arguments{
  \item{file.grid}{CLM file name with extension *.clm}
  \item{endian}{endianess of the file}
  \item{\dots}{Further arguments (currently not used).}
}
\details{No details.}

\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}



\seealso{\code{\link{ReadCLM}}}

