\name{ReadPRO}
\alias{ReadPRO}
\title{Read *.pro files as produced from GENOUD}
\description{The function is used within OptimizeLPJgenoud}
\usage{ReadPRO(files.pro)}
\arguments{
  \item{files.pro}{file names (*.pro) of genoud optimization results.}
}
\details{No details.}
\value{The function returns a data.frame with number of individual, cost and parameer values}
\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





