\name{StandardError}
\alias{StandardError}
\title{Compute standard errors from a variance-covariance matrix}
\description{SE = sqrt(diag(vc) * cost^2 / (nobs - npar))}
\usage{StandardError(vc, nobs, cost)}
\arguments{
  \item{vc}{variance-covariance matrix}
  \item{nobs}{number of observations}
  \item{cost}{cost function value}
}
\details{No details.}

\references{No reference.	}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





