\name{Turnover}
\alias{Turnover}
\title{Calculate turnover time from stock and flux}
\description{Calculates turnover times.}
\usage{Turnover(stock, flux, ...)}
\arguments{
  \item{stock}{stock, e.g. biomass}
  \item{flux}{flux, e.g. NPP}
  \item{\dots}{further arguments (currently not used)}
}
\details{turnover = stock / flux}

\references{No reference.}
\author{Matthias Forkel <matthias.forkel@geo.tuwien.ac.at> [aut, cre]}





